+++ src/works/Works.ts
@@ -1,5 +1,5 @@
 import { isNew } from '../utils';
-import { Citation, CitationMap, toCitationMap, Category } from './bibliography';
+import { Citation, CitationMap, toCitationMap, Genre } from './citation';
 
 const Text = (text: string) =>
   text.replace(
@@ -11,13 +11,10 @@ const Item = (text: string) => `<li>${Text(text)}</li>`;
 
 const HilightedItem = (text: string) => `<string>${Item(text)}</strong>`;
 
-const List = (
-  citations: Citation[] | undefined,
-  category: Category
-): string => {
+const List = (citations: Citation[] | undefined, genre: Genre): string => {
   if (!citations || citations.length === 0) return '';
   return `
-    <h3>${category}</h3>
+    <h3>${genre}</h3>
     <ol>
       ${citations.map((citation) => Item(citation._bibliographyText)).join('')}
     </ol>`;
@@ -30,6 +27,6 @@ export const Works = (citations: Citation[]): string => {
   <section class="ly_cont">
   <div class="bl_text">
     <h2>業績一覧</h2>
-    ${Category.map((c) => List(citationMap.get(c), c)).join('')}
+    ${Genre.map((g) => List(citationMap.get(g), g)).join('')}
   </div>`;
 };
