+++ src/works/index.ts
@@ -1,14 +1,11 @@
 import { Data } from 'csl-json';
 import { renderRoot } from '../render';
 import { fetcher } from '../utils';
-import { Citation } from './bibliography';
+import { Citation } from './citation';
 import { Works } from './Works';
 
 const jsonPath = './works.json';
 
-async function index() {
-  const works = await fetcher<Citation[]>(jsonPath);
-  renderRoot(Works(works));
-}
-
-index();
+(async function index() {
+  renderRoot(Works(await fetcher<Citation[]>(jsonPath)));
+})();
