+++ works.json
@@ -29,8 +29,8 @@
     "page": "89-97",
     "ISSN": "0289-825X",
     "URL": null,
-    "note": null,
-    "_bibliographyText": "京念屋隆史「考えるとき話しているのか——フッサール「独白」概念をめぐって」『現象学年報』(37): pp. 89–97, 2021."
+    "note": "[要旨](https://github.com/kyonenya/kyonenya.github.io/blob/main/docs/211126-Am_I_speaking_when_I_think.md)",
+    "_bibliographyText": "京念屋隆史「考えるとき話しているのか——フッサール「独白」概念をめぐって」『現象学年報』(37): pp. 89–97, 2021. [要旨](https://github.com/kyonenya/kyonenya.github.io/blob/main/docs/211126-Am_I_speaking_when_I_think.md)"
   },
   {
     "id": 4,
